fix a bunch

+++ rust/lance-linalg/src/simd/f16.c
@@ -20,14 +20,14 @@
 #endif // __X86_64__
 
 /// Works on NEON + FP16 or AVX512FP16
-_Float16 norm_l2_f16(const _Float16 *data, uint32_t dimension) {
+float norm_l2_f16(const _Float16 *data, uint32_t dimension) {
   _Float16 sum = 0;
 
 #pragma clang loop unroll(enable) vectorize(enable) interleave(enable)
   for (uint32_t i = 0; i < dimension; i++) {
     sum += data[i] * data[i];
   }
-  return sum;
+  return (float) sum;
 }
 
 /// @brief Dot product of two f16 vectors.
@@ -35,17 +35,17 @@ _Float16 norm_l2_f16(const _Float16 *data, uint32_t dimension) {
 /// @param y A f16 vector
 /// @param dimension The dimension of the vectors
 /// @return The dot product of the two vectors.
-_Float16 dot_f16(const _Float16 *x, const _Float16 *y, uint32_t dimension) {
+float dot_f16(const _Float16 *x, const _Float16 *y, uint32_t dimension) {
   _Float16 sum = 0;
 
 #pragma clang loop unroll(enable) interleave(enable) vectorize(enable)
   for (uint32_t i = 0; i < dimension; i++) {
     sum += x[i] * y[i];
   }
-  return sum;
+  return (float) sum;
 }
 
-_Float16 l2_f16(const _Float16 *x, const _Float16 *y, uint32_t dimension) {
+float l2_f16(const _Float16 *x, const _Float16 *y, uint32_t dimension) {
   _Float16 sum = 0.0;
 
 #pragma clang loop unroll(enable) interleave(enable) vectorize_width(32)
@@ -53,5 +53,5 @@ _Float16 l2_f16(const _Float16 *x, const _Float16 *y, uint32_t dimension) {
     _Float16 s = x[i] - y[i];
     sum += s * s;
   }
-  return sum;
+  return (float) sum;
 }
